breaking up the CSVGateway with a handler
